dateStrの共通化
