+++ app/javascript/components/modules/comments/comment.vue
@@ -76,22 +76,6 @@ import CommentForm from "./comment_form.vue";
 
 Vue.use(Lightbox);
 
-const dateFormatter = Intl.DateTimeFormat('ja-JP', {
-  month: 'narrow',
-  day: 'numeric' ,
-  hour12: false,
-  hour: '2-digit',
-  minute: '2-digit'
-});
-const dateFormatterWithYear = Intl.DateTimeFormat('ja-JP', {
-  year: 'numeric',
-  month: 'narrow',
-  day: 'numeric' ,
-  hour12: false,
-  hour: '2-digit',
-  minute: '2-digit'
-});
-
 export default {
   name: "comment",
   props: {
@@ -141,12 +125,6 @@ export default {
     toggleMuted() {
       this.muted = !this.muted;
       if(this.muted) this.showMuteDialog();
-    },
-    dateStr(date) {
-      if(date.getFullYear() != new Date().getFullYear()){
-        return dateFormatterWithYear.format(date);
-      }
-      return dateFormatter.format(date);
     }
   }
 }
